[mod/#140] dto 분리
